Added another FAQ for beginners with a specific error
Since this seems intended to people who don't want to commit to TS, it seems like this issue might be a bit strange and it took me a bit of fumbling around to find out this simple solution in the end. So I added a FAQ question that I think might be useful for future people who might encounter this error.